% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca.R
\name{get_local_loadings}
\alias{get_local_loadings}
\title{Extract loadings for a specific location}
\usage{
get_local_loadings(object, location, n_components = NULL)
}
\arguments{
\item{object}{A gwpca object}

\item{location}{Integer index or coordinate vector}

\item{n_components}{Number of components to extract}
}
\value{
Matrix of loadings
}
\description{
Extract loadings for a specific location
}
