% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca.R
\name{get_local_variance}
\alias{get_local_variance}
\title{Extract variance explained at each location}
\usage{
get_local_variance(object, cumulative = FALSE)
}
\arguments{
\item{object}{A gwpca object}

\item{cumulative}{Logical, whether to return cumulative variance}
}
\value{
Matrix of variance explained
}
\description{
Extract variance explained at each location
}
