% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca.R
\name{predict.gwpca}
\alias{predict.gwpca}
\title{Predict method for gwpca objects}
\usage{
\method{predict}{gwpca}(object, newdata, newcoords, ...)
}
\arguments{
\item{object}{A gwpca object}

\item{newdata}{Matrix of new data points}

\item{newcoords}{Matrix of new coordinates}

\item{...}{Additional arguments (ignored)}
}
\value{
Matrix of predicted scores
}
\description{
Predict method for gwpca objects
}
