% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca.R
\name{print.gwpca}
\alias{print.gwpca}
\title{Print method for gwpca objects}
\usage{
\method{print}{gwpca}(x, ...)
}
\arguments{
\item{x}{A gwpca object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for gwpca objects
}
