% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca.R
\name{print.summary.gwpca}
\alias{print.summary.gwpca}
\title{Print method for summary.gwpca objects}
\usage{
\method{print}{summary.gwpca}(x, ...)
}
\arguments{
\item{x}{A summary.gwpca object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for summary.gwpca objects
}
