% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca.R
\name{summary.gwpca}
\alias{summary.gwpca}
\title{Summary method for gwpca objects}
\usage{
\method{summary}{gwpca}(object, ...)
}
\arguments{
\item{object}{A gwpca object}

\item{...}{Additional arguments (ignored)}
}
\value{
A summary.gwpca object
}
\description{
Summary method for gwpca objects
}
